use multi-stage build in Dockerfile to halve size
This commit proposes using a multi-stage Dockerfile. In the first stage,
we use the official python:3.8 image, which includes a C compilers. We
install the project and its dependencies. We create a dedicated virtual
environment for the project. Then in the second stage, we copy over the
project files and the python virtual environment, and we install
libopenslide0, which is required for openslide-python runtime.

This results in a more complex Dockerfile, but the resulting image is
half as big as the original (671 MB from about 1.3GB). I am also happy
to maintain this Dockerfile and answer any questions about it.