install wheel + move apt install up
In the second stage, we can move the apt install step up, because it
does not rely on any artifacts from the first build stage. This should
speed up image builds slightly, when using buildkit. Buildkit builds
different stages in parallel when it can (i.e., it builds the parts of
the stage that do not rely on previous stages).