expanduser in basepath and outdir
Closes https://github.com/choosehappy/HistoQC/issues/197

When using a path with a tilde `~` on unix-like systems, the path is
interpreted as a literal `./~`, meaning a directory named `~` in the
current working directory. Users typically do not want this and would
probably be surprised by this behavior. I know I was surprised when I
ran histoqc and then saw a new directory named `~`. I removed it very
carefully, trying to be sure I didn't trash my home directory :)

This commit fixes the issue by expanding the tilde into the path to the
home directory. If a path does not begin with a tilde, the path is
unchanged.